LayoutTests/imported/w3c/web-platform-tests/css/css-color/animation/color-composition-expected.txt


FAIL Compositing: property <color> underlying [rgb(50, 50, 50)] from add [rgb(10, 10, 10)] to replace [rgb(30, 30, 30)] at (0) should be [rgb(60, 60, 60)] assert_equals: expected "rgb ( 60 , 60 , 60 ) " but got "rgb ( 10 , 10 , 10 ) "
FAIL Compositing: property <color> underlying [rgb(50, 50, 50)] from add [rgb(10, 10, 10)] to replace [rgb(30, 30, 30)] at (0.2) should be [rgb(54, 54, 54)] assert_equals: expected "rgb ( 54 , 54 , 54 ) " but got "rgb ( 14 , 14 , 14 ) "
PASS Compositing: property <color> underlying [rgb(50, 50, 50)] from add [rgb(10, 10, 10)] to replace [rgb(30, 30, 30)] at (1) should be [rgb(30, 30, 30)]
FAIL Compositing: property <color> underlying [rgb(50, 50, 50)] from add [rgb(10, 10, 10)] to replace [rgb(30, 30, 30)] at (1.2) should be [rgb(24, 24, 24)] assert_equals: expected "rgb ( 24 , 24 , 24 ) " but got "rgb ( 34 , 34 , 34 ) "
FAIL Compositing: property <color> underlying [rgb(50, 50, 50)] from add [rgb(10, 10, 10)] to replace [rgb(30, 30, 30)] at (1.5) should be [rgb(15, 15, 15)] assert_equals: expected "rgb ( 15 , 15 , 15 ) " but got "rgb ( 40 , 40 , 40 ) "
FAIL Compositing: property <color> underlying [rgb(60, 60, 60)] from add [rgb(0, 0, 0)] to replace [rgb(50, 50, 50)] at (0) should be [rgb(60, 60, 60)] assert_equals: expected "rgb ( 60 , 60 , 60 ) " but got "rgb ( 0 , 0 , 0 ) "
FAIL Compositing: property <color> underlying [rgb(60, 60, 60)] from add [rgb(0, 0, 0)] to replace [rgb(50, 50, 50)] at (0.5) should be [rgb(55, 55, 55)] assert_equals: expected "rgb ( 55 , 55 , 55 ) " but got "rgb ( 25 , 25 , 25 ) "
PASS Compositing: property <color> underlying [rgb(60, 60, 60)] from add [rgb(0, 0, 0)] to replace [rgb(50, 50, 50)] at (1) should be [rgb(50, 50, 50)]
FAIL Compositing: property <color> underlying [rgb(60, 60, 60)] from add [rgb(0, 0, 0)] to replace [rgb(50, 50, 50)] at (1.2) should be [rgb(48, 48, 48)] assert_equals: expected "rgb ( 48 , 48 , 48 ) " but got "rgb ( 60 , 60 , 60 ) "
FAIL Compositing: property <color> underlying [rgb(60, 60, 60)] from add [rgb(0, 0, 0)] to replace [rgb(50, 50, 50)] at (1.5) should be [rgb(45, 45, 45)] assert_equals: expected "rgb ( 45 , 45 , 45 ) " but got "rgb ( 75 , 75 , 75 ) "
